Settings refactoring: use CtxSettings, remove BasicSettings
