+++ core/include/gnuradio-4.0/Settings.hpp
@@ -18,7 +18,17 @@
 #include <gnuradio-4.0/meta/formatter.hpp>
 #include <gnuradio-4.0/reflection.hpp>
 
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wformat-nonliteral"
+#include <fmt/chrono.h>
+#pragma GCC diagnostic pop
+
 namespace gr {
+
+namespace detail {
+template<typename T>
+concept HasBaseType = requires { typename std::remove_cvref_t<T>::base_t; };
+}; // namespace detail
 
 namespace settings {
 
@@ -52,6 +62,16 @@ inline constexpr uint64_t convertTimePointToUint64Ns(const std::chrono::time_poi
     const auto ns = std::chrono::duration_cast<std::chrono::nanoseconds>(tp.time_since_epoch()).count();
     return static_cast<uint64_t>(ns);
 }
+
+static auto nullMatchPred = [](auto, auto, auto) { return std::nullopt; };
+
+/**
+ * Policy for handling Settings Auto Update when `context` is not found in storedParameters.
+ */
+enum class AutoUpdatePolicy {
+    Ignore,     // do nothing when context is not in storedParameters
+    AddToStored // Add the new context to storedParameters with empty parameters and apply any new changes present in the Tag
+};
 
 // pmtv::pmt comparison is needed to use it as a key of std::map
 struct PMTCompare {
@@ -132,106 +152,14 @@ concept HasSettingsResetCallback = requires(TBlock* block) {
     { block->reset() };
 };
 
-template<typename T>
-concept SettingsLike = requires(T t, std::span<const std::string> parameter_keys, const std::string& parameter_key, const property_map& parameters, SettingsCtx ctx, const Tag& tag) {
-    /**
-     * @brief returns if there are stages settings that haven't been applied yet.
-     */
-    { t.changed() } -> std::same_as<bool>;
-
-    /**
-     * @brief stages new key-value pairs that shall replace the block field-based settings.
-     * N.B. settings become only active after executing 'applyStagedParameters()' (usually done early on in the 'Block::work()' function)
-     * @return key-value pairs that could not be set
-     */
-    { t.set(parameters, ctx) } -> std::same_as<property_map>;
-    { t.set(parameters) } -> std::same_as<property_map>;
-
-    /**
-     * @brief updates parameters based on block input tags for those with keys stored in `autoUpdateParameters()`
-     * Parameter changes to down-stream blocks is controlled via `autoForwardParameters()`
-     */
-    { t.autoUpdate(tag) } -> std::same_as<void>;
-
-    /**
-     * @brief return all available active block settings as key-value pairs
-     */
-    { t.get() } -> std::same_as<property_map>;
-
-    /**
-     * @brief return all available active block settings as key-value pairs for multiple keys
-     */
-    { t.get(parameter_keys) } -> std::same_as<property_map>;
-
-    /**
-     * @brief return available active block setting as key-value pair for a single key
-     */
-    { t.get(parameter_key) } -> std::same_as<std::optional<pmtv::pmt>>;
-
-    /**
-     * @brief return all available stored block settings for default SettingsCtx as key-value pairs
-     */
-    { t.getStored() } -> std::same_as<property_map>;
-
-    /**
-     * @brief return all available stored block settings as key-value pairs for multiple keys
-     */
-    { t.getStored(parameter_keys) } -> std::same_as<property_map>;
-    { t.getStored(parameter_keys, ctx) } -> std::same_as<property_map>;
-
-    /**
-     * @brief return available stored block setting as key-value pair for a single key
-     */
-    { t.getStored(parameter_key) } -> std::same_as<std::optional<pmtv::pmt>>;
-    { t.getStored(parameter_key, ctx) } -> std::same_as<std::optional<pmtv::pmt>>;
-
-    /**
-     * @brief return number of all stored parameters
-     */
-    { t.getNStoredParameters() } -> std::same_as<gr::Size_t>;
-
-    /**
-     * @brief return _storedParameters
-     */
-    { t.getStoredAll() } -> std::same_as<std::map<pmtv::pmt, std::vector<std::pair<SettingsCtx, property_map>>, settings::PMTCompare>>;
-
-    /**
-     * @brief returns the staged/not-yet-applied new parameters
-     */
-    { t.stagedParameters() } -> std::same_as<const property_map>;
-
-    /**
-     * @brief synchronise map-based with actual block field-based settings
-     */
-    { t.applyStagedParameters() } -> std::same_as<ApplyStagedParametersResult>;
-    { t.applyStagedParameters(0ULL) } -> std::same_as<ApplyStagedParametersResult>;
-
-    /**
-     * @brief synchronises the map-based with the block's field-based parameters
-     * (N.B. usually called after the staged parameters have been synchronised)
-     */
-    { t.updateActiveParameters() } -> std::same_as<void>;
-};
-
 struct SettingsBase {
-    std::atomic_bool _changed{false};
-
     virtual ~SettingsBase() = default;
 
-    void swap(SettingsBase& other) noexcept {
-        if (this == &other) {
-            return;
-        }
-        bool temp = _changed;
-        // avoid CAS-loop since this called only during initialisation where there is no concurrent access possible.
-        std::atomic_store_explicit(&_changed, std::atomic_load_explicit(&other._changed, std::memory_order_acquire), std::memory_order_release);
-        other._changed = temp;
-    }
-
     /**
-     * @brief returns if there are stages settings that haven't been applied yet.
+     * @brief returns if there are stored settings that haven't been applied yet.
      */
-    [[nodiscard]] bool changed() const noexcept { return _changed; }
+    [[nodiscard]] virtual bool changed() const noexcept    = 0;
+    virtual void               setChanged(bool b) noexcept = 0;
 
     /**
      * @brief stages new key-value pairs that shall replace the block field-based settings.
@@ -300,31 +228,41 @@ struct SettingsBase {
      * (N.B. usually called after the staged parameters have been synchronised)
      */
     virtual void updateActiveParameters() noexcept = 0;
-};
-
-namespace detail {
-template<typename T>
-concept HasBaseType = requires { typename std::remove_cvref_t<T>::base_t; };
-}; // namespace detail
+}; // struct SettingsBase
 
 template<typename TBlock>
-class BasicSettings : public SettingsBase {
-    TBlock*                            _block = nullptr;
-    mutable std::mutex                 _lock{};
-    property_map                       _active{}; // copy of class field settings as pmt-style map
-    property_map                       _staged{}; // parameters to become active before the next work() call
-    std::set<std::string, std::less<>> _auto_update{};
-    std::set<std::string, std::less<>> _auto_forward{};
-    property_map                       _default_settings{};
+class CtxSettings : public SettingsBase {
+    /**
+     * A predicate for matching two contexts
+     * The third "attempt" parameter indicates the current round of matching being done.
+     * This is useful for hierarchical matching schemes,
+     * e.g. in the first round the predicate could look for almost exact matches only,
+     * then in a a second round (attempt=1) it could be more forgiving, given that there are no exact matches available.
+     *
+     * The predicate will be called until it returns "true" (a match is found), or until it returns std::nullopt,
+     * which indicates that no matches were found and there is no chance of matching anything in a further round.
+     */
+    using MatchPredicate = std::function<std::optional<bool>(const pmtv::pmt&, const pmtv::pmt&, std::size_t)>;
+
+    TBlock*            _block = nullptr;
+    std::atomic_bool   _changed{false};
+    mutable std::mutex _mutex{};
+    property_map       _activeParameters{};
+    // key is SettingsCtx.context, value: queue of parameters with the same SettingsCtx.context but for different time
+    mutable std::map<pmtv::pmt, std::vector<std::pair<SettingsCtx, property_map>>, settings::PMTCompare> _storedParameters{};
+    property_map                                                                                         _defaultParameters{};
+    std::set<std::string, std::less<>>                                                                   _allWritableMembers{};   // all `isWritableMember` class members
+    std::map<pmtv::pmt, std::set<std::string, std::less<>>, settings::PMTCompare>                        _autoUpdateParameters{}; // for each SettingsCtx.context auto updated members are store separately
+    std::set<std::string, std::less<>>                                                                   _autoForwardParameters{};
+    MatchPredicate                                                                                       _matchPred        = settings::nullMatchPred;
+    pmtv::pmt                                                                                            _activeCtx        = "";
+    settings::AutoUpdatePolicy                                                                           _autoUpdatePolicy = settings::AutoUpdatePolicy::AddToStored;
 
 public:
-    BasicSettings()  = delete;
-    ~BasicSettings() = default;
-
-    explicit constexpr BasicSettings(TBlock& block) noexcept : SettingsBase(), _block(&block) {
+    explicit CtxSettings(TBlock& block, MatchPredicate matchPred = settings::nullMatchPred) noexcept : SettingsBase(), _block(&block), _matchPred(matchPred) {
         if constexpr (requires { &TBlock::settingsChanged; }) { // if settingsChanged is defined
             static_assert(HasSettingsChangedCallback<TBlock>, "if provided, settingsChanged must have either a `(const property_map& old, property_map& new, property_map& fwd)`"
-                                                              "or `(const property_map& old, property_map& new)` parameter signatures.");
+                                                              "or `(const property_map& old, property_map& new)` paremeter signatures.");
         }
 
         if constexpr (requires { &TBlock::reset; }) { // if reset is defined
@@ -332,7 +270,6 @@ public:
         }
 
         if constexpr (refl::is_reflectable<TBlock>()) {
-            // register block-global description
             constexpr bool hasMetaInfo = requires(TBlock t) {
                 {
                     unwrap_if_wrapped_t<decltype(t.meta_information)> {}
@@ -345,8 +282,8 @@ public:
             }
 
             // handle meta-information for UI and other non-processing-related purposes
-            auto processOneMember = [this]<typename Member>(Member member) {
-                using RawType         = std::remove_cvref_t<decltype(member(*_block))>;
+            auto processOneMember = [&]<typename TFieldMeta>(TFieldMeta member) {
+                using RawType         = std::remove_cvref_t<typename TFieldMeta::value_type>;
                 using Type            = unwrap_if_wrapped_t<RawType>;
                 const auto memberName = std::string(get_display_name(member));
 
@@ -357,68 +294,106 @@ public:
                     _block->meta_information.value[memberName + "::visible"]       = RawType::visible();
                 }
 
-                // detect whether field has one of the kDefaultTags signature
+                // detect whether field has one of the DEFAULT_TAGS signature
                 if constexpr (settings::isWritableMember<Type>(member)) {
                     if constexpr (std::ranges::find(gr::tag::kDefaultTags, std::string_view(get_display_name_const(member).c_str())) != gr::tag::kDefaultTags.cend()) {
-                        _auto_forward.emplace(memberName);
+                        _autoForwardParameters.emplace(memberName);
                     }
-                    _auto_update.emplace(memberName);
+                    _allWritableMembers.emplace(memberName);
                 }
             };
             processMembers<TBlock>(processOneMember);
         }
+        addStoredParameters(property_map(), SettingsCtx(0ULL, ""));
     }
 
-    constexpr BasicSettings(const BasicSettings& other) noexcept : SettingsBase(other) {
-        BasicSettings temp(other);
-        swap(temp);
+    CtxSettings(const CtxSettings& other) {
+        std::scoped_lock lock(_mutex, other._mutex);
+        copyFrom(other);
     }
 
-    constexpr BasicSettings(BasicSettings&& other) noexcept : SettingsBase(std::move(other)) {
-        BasicSettings temp(std::move(other));
-        swap(temp);
+    CtxSettings(CtxSettings&& other) noexcept {
+        std::scoped_lock lock(_mutex, other._mutex);
+        moveFrom(other);
     }
 
-    BasicSettings& operator=(const BasicSettings& other) noexcept {
-        swap(other);
-        return *this;
-    }
-
-    BasicSettings& operator=(BasicSettings&& other) noexcept {
-        BasicSettings temp(std::move(other));
-        swap(temp);
-        return *this;
-    }
-
-    void swap(BasicSettings& other) noexcept {
+    CtxSettings& operator=(const CtxSettings& other) noexcept {
         if (this == &other) {
-            return;
+            return *this;
         }
-        SettingsBase::swap(other);
-        std::swap(_block, other._block);
-        std::scoped_lock lock(_lock, other._lock);
-        std::swap(_active, other._active);
-        std::swap(_staged, other._staged);
-        std::swap(_auto_update, other._auto_update);
-        std::swap(_auto_forward, other._auto_forward);
+
+        std::scoped_lock lock(_mutex, other._mutex);
+        copyFrom(other);
+        return *this;
     }
 
-    [[nodiscard]] property_map set(const property_map& parameters, SettingsCtx = {}) override {
+    CtxSettings& operator=(CtxSettings&& other) noexcept {
+        if (this == &other) {
+            return *this;
+        }
+        std::scoped_lock lock(_mutex, other._mutex);
+        moveFrom(other);
+        return *this;
+    }
+
+private:
+    void copyFrom(const CtxSettings& other) {
+        _block = other._block;
+        std::atomic_store_explicit(&_changed, std::atomic_load_explicit(&other._changed, std::memory_order_acquire), std::memory_order_release);
+        _activeParameters      = other._activeParameters;
+        _storedParameters      = other._storedParameters;
+        _defaultParameters     = other._defaultParameters;
+        _allWritableMembers    = other._allWritableMembers;
+        _autoUpdateParameters  = other._autoUpdateParameters;
+        _autoForwardParameters = other._autoForwardParameters;
+        _matchPred             = other._matchPred;
+        _activeCtx             = other._activeCtx;
+        _autoUpdatePolicy      = other._autoUpdatePolicy;
+    }
+
+    void moveFrom(CtxSettings& other) noexcept {
+        _block = std::exchange(other._block, nullptr);
+        std::atomic_store_explicit(&_changed, std::atomic_load_explicit(&other._changed, std::memory_order_acquire), std::memory_order_release);
+        _activeParameters      = std::move(other._activeParameters);
+        _storedParameters      = std::move(other._storedParameters);
+        _defaultParameters     = std::move(other._defaultParameters);
+        _allWritableMembers    = std::move(other._allWritableMembers);
+        _autoUpdateParameters  = std::move(other._autoUpdateParameters);
+        _autoForwardParameters = std::move(other._autoForwardParameters);
+        _matchPred             = std::exchange(other._matchPred, settings::nullMatchPred);
+        _activeCtx             = std::exchange(other._activeCtx, "");
+        _autoUpdatePolicy      = std::exchange(other._autoUpdatePolicy, settings::AutoUpdatePolicy::AddToStored);
+    }
+
+public:
+    [[nodiscard]] bool changed() const noexcept override { return _changed; }
+
+    void setChanged(bool b) noexcept override { _changed.store(b); }
+
+    [[nodiscard]] property_map set(const property_map& parameters, SettingsCtx ctx = {}) override {
         property_map ret;
         if constexpr (refl::is_reflectable<TBlock>()) {
-            std::lock_guard lg(_lock);
+            std::lock_guard lg(_mutex);
+            if (ctx.time == 0ULL) {
+                ctx.time = settings::convertTimePointToUint64Ns(std::chrono::system_clock::now());
+            }
+            property_map newParameters = getBestMatchStoredParameters(ctx);
+            if (_autoUpdateParameters.find(ctx.context) == _autoUpdateParameters.end()) {
+                _autoUpdateParameters[ctx.context] = _allWritableMembers;
+            }
+
             for (const auto& [key, value] : parameters) {
                 bool isSet            = false;
-                auto processOneMember = [&, this](auto member) {
-                    using Type = unwrap_if_wrapped_t<std::remove_cvref_t<decltype(member(*_block))>>;
+                auto processOneMember = [&, this]<typename TFieldMeta>(TFieldMeta member) {
+                    using Type = unwrap_if_wrapped_t<std::remove_cvref_t<typename TFieldMeta::value_type>>;
                     if constexpr (settings::isWritableMember<Type>(member)) {
                         const auto fieldName = std::string_view(get_display_name(member));
                         if (fieldName == key && std::holds_alternative<Type>(value)) {
-                            if (_auto_update.contains(key)) {
-                                _auto_update.erase(key);
+                            if (_autoUpdateParameters[ctx.context].contains(key)) {
+                                _autoUpdateParameters[ctx.context].erase(key);
                             }
-                            _staged.insert_or_assign(key, value);
-                            SettingsBase::_changed.store(true);
+                            newParameters.insert_or_assign(key, value);
+                            _changed.store(true);
                             isSet = true;
                         }
                         if (fieldName == key && !std::holds_alternative<Type>(value)) {
@@ -435,9 +410,10 @@ public:
                     ret.insert_or_assign(key, pmtv::pmt(value));
                 }
             }
+            addStoredParameters(newParameters, ctx);
         }
 
-        // copy items that could not be matched to the block's meta_information map (if available)
+        // copy items that could not be matched to the node's meta_information map (if available)
         if constexpr (requires(TBlock t) {
                           {
                               unwrap_if_wrapped_t<decltype(t.meta_information)> {}
@@ -449,10 +425,12 @@ public:
         return ret; // N.B. returns those <key:value> parameters that could not be set
     }
 
-    void storeDefaults() override { this->storeDefaultSettings(_default_settings); }
+    void storeDefaults() override { this->storeDefaultSettings(_defaultParameters); }
 
     void resetDefaults() override {
-        _staged     = _default_settings;
+        _storedParameters.clear();
+        _autoUpdateParameters.clear();
+        addStoredParameters(_defaultParameters, SettingsCtx(settings::convertTimePointToUint64Ns(std::chrono::system_clock::now()), ""));
         std::ignore = applyStagedParameters();
         if constexpr (HasSettingsResetCallback<TBlock>) {
             _block->reset();
@@ -460,120 +438,147 @@ public:
     }
 
     void autoUpdate(const Tag& tag) override {
-        const auto& parameters = tag.map;
         if constexpr (refl::is_reflectable<TBlock>()) {
+            SettingsCtx ctx = createSettingsCtxFromTag(tag);
+
+            auto bestMatchCtx = findBestMatchCtx(ctx.context);
+            if (bestMatchCtx == std::nullopt) {
+                if (_autoUpdatePolicy == settings::AutoUpdatePolicy::AddToStored) {
+                    bestMatchCtx = ctx.context;
+                } else if (_autoUpdatePolicy == settings::AutoUpdatePolicy::Ignore) {
+                    return;
+                }
+            }
+
+            const auto  found                = _autoUpdateParameters.find(bestMatchCtx.value());
+            const auto& autoUpdateParameters = found == _autoUpdateParameters.end() ? _allWritableMembers : found->second;
+
+            property_map        newParameters = getBestMatchStoredParameters(ctx);
+            const property_map& parameters    = tag.map;
+            bool                wasChanged    = false;
             for (const auto& [key, value] : parameters) {
-                auto processOneMember = [&](auto member) {
-                    using Type = unwrap_if_wrapped_t<std::remove_cvref_t<decltype(member(*_block))>>;
+                auto processOneMember = [&]<typename TFieldMeta>(TFieldMeta member) {
+                    using Type = unwrap_if_wrapped_t<std::remove_cvref_t<typename TFieldMeta::value_type>>;
                     if constexpr (settings::isWritableMember<Type>(member)) {
-                        if (_auto_update.contains(key) && std::string(get_display_name(member)) == key && std::holds_alternative<Type>(value)) {
-                            _staged.insert_or_assign(key, value);
-                            SettingsBase::_changed.store(true);
+                        if (std::string_view(get_display_name(member)) == key && autoUpdateParameters.contains(key) && std::holds_alternative<Type>(value)) {
+                            newParameters.insert_or_assign(key, value);
+                            wasChanged = true;
                         }
                     }
                 };
                 processMembers<TBlock>(processOneMember);
             }
-        }
-    }
-
-    [[nodiscard]] const property_map stagedParameters(SettingsCtx = {}) const noexcept override {
-        std::lock_guard lg(_lock);
-        return _staged;
-    }
-
-    [[nodiscard]] property_map get(std::span<const std::string> parameter_keys = {}) const noexcept override {
-        std::lock_guard lg(_lock);
-        if (parameter_keys.empty()) {
-            return _active;
-        }
-        property_map ret;
-        for (const auto& key : parameter_keys) {
-            if (_active.contains(key)) {
-                ret.insert_or_assign(key, _active.at(key));
+            if (wasChanged || isContextPresentInTag(tag)) {
+                addStoredParameters(newParameters, SettingsCtx(ctx.time, bestMatchCtx.value()));
+                _activeCtx = bestMatchCtx.value();
+                _changed.store(true);
             }
         }
-        return ret;
     }
 
-    [[nodiscard]] std::optional<pmtv::pmt> get(const std::string& parameter_key) const noexcept override {
-        if constexpr (refl::is_reflectable<TBlock>()) {
-            std::lock_guard lg(_lock);
-
-            if (_active.contains(parameter_key)) {
-                return {_active.at(parameter_key)};
-            }
-        }
-
-        return std::nullopt;
+    [[nodiscard]] const property_map stagedParameters(SettingsCtx ctx = {}) const noexcept override {
+        std::lock_guard lg(_mutex);
+        return hasStoredParameters(ctx) ? getBestMatchStoredParameters(ctx) : property_map();
     }
 
-    [[nodiscard]] property_map getStored(std::span<const std::string> parameterKeys = {}, SettingsCtx = {}) const noexcept override {
-        std::lock_guard lg(_lock);
+    [[nodiscard]] property_map get(std::span<const std::string> parameterKeys = {}) const noexcept override {
+        std::lock_guard lg(_mutex);
         if (parameterKeys.empty()) {
-            return _staged;
+            return _activeParameters;
         }
         property_map ret;
         for (const auto& key : parameterKeys) {
-            if (_staged.contains(key)) {
-                ret.insert_or_assign(key, _staged.at(key));
+            if (_activeParameters.contains(key)) {
+                ret.insert_or_assign(key, _activeParameters.at(key));
             }
         }
         return ret;
     }
 
-    [[nodiscard]] std::optional<pmtv::pmt> getStored(const std::string& parameter_key, SettingsCtx = {}) const noexcept override {
-        auto res = getStored(std::array<std::string, 1>({parameter_key}));
-        if (res.contains(parameter_key)) {
-            return res.at(parameter_key);
+    [[nodiscard]] std::optional<pmtv::pmt> get(const std::string& parameterKey) const noexcept override {
+        auto res = get(std::array<std::string, 1>({parameterKey}));
+        if (res.contains(parameterKey)) {
+            return res.at(parameterKey);
         } else {
             return std::nullopt;
         }
     }
 
-    [[nodiscard]] gr::Size_t getNStoredParameters() const noexcept override { return 1; } // Implemented only for compatibility
+    [[nodiscard]] property_map getStored(std::span<const std::string> parameterKeys = {}, SettingsCtx ctx = {}) const noexcept override {
+        std::lock_guard     lg(_mutex);
+        const property_map& allBestMatchParameters = this->getBestMatchStoredParameters(ctx);
 
-    [[nodiscard]] std::map<pmtv::pmt, std::vector<std::pair<SettingsCtx, property_map>>, settings::PMTCompare> getStoredAll() const noexcept override { return {}; } // Implemented only for compatibility
+        if (parameterKeys.empty()) {
+            return allBestMatchParameters;
+        }
+        property_map ret;
+        for (const auto& key : parameterKeys) {
+            if (allBestMatchParameters.contains(key)) {
+                ret.insert_or_assign(key, allBestMatchParameters.at(key));
+            }
+        }
+        return ret;
+    }
 
-    [[nodiscard]] std::set<std::string, std::less<>> autoUpdateParameters(SettingsCtx = {}) noexcept override { return _auto_update; }
+    [[nodiscard]] std::optional<pmtv::pmt> getStored(const std::string& parameterKey, SettingsCtx ctx = {}) const noexcept override {
+        auto res = getStored(std::array<std::string, 1>({parameterKey}), ctx);
+        if (res.contains(parameterKey)) {
+            return res.at(parameterKey);
+        } else {
+            return std::nullopt;
+        }
+    }
 
-    [[nodiscard]] std::set<std::string, std::less<>>& autoForwardParameters() noexcept override { return _auto_forward; }
+    [[nodiscard]] gr::Size_t getNStoredParameters() const noexcept override {
+        std::lock_guard lg(_mutex);
+        gr::Size_t      nParameters{0};
+        for (const auto& stored : _storedParameters) {
+            nParameters += static_cast<gr::Size_t>(stored.second.size());
+        }
+        return nParameters;
+    }
 
-    /**
-     * @brief synchronise map-based with actual block field-based settings
-     * returns a structure containing three maps:
-     *  - forwardParameters -- map with key-value tags that should be forwarded
-     *    to dependent/child blocks.
-     *  - appliedParameters -- map with peoperties that were successfully set
-     */
-    [[nodiscard]] ApplyStagedParametersResult applyStagedParameters(std::uint64_t = 0ULL) override {
+    [[nodiscard]] std::map<pmtv::pmt, std::vector<std::pair<SettingsCtx, property_map>>, settings::PMTCompare> getStoredAll() const noexcept override { return _storedParameters; }
+
+    [[nodiscard]] std::set<std::string, std::less<>> autoUpdateParameters(SettingsCtx ctx = {}) noexcept override {
+        auto bestMatchCtx = findBestMatchCtx(ctx.context);
+        return bestMatchCtx == std::nullopt ? std::set<std::string, std::less<>>() : _autoUpdateParameters[bestMatchCtx.value()];
+    }
+
+    [[nodiscard]] std::set<std::string, std::less<>>& autoForwardParameters() noexcept override { return _autoForwardParameters; }
+
+    [[nodiscard]] ApplyStagedParametersResult applyStagedParameters(std::uint64_t currentTime = 0ULL) override {
         ApplyStagedParametersResult result;
         if constexpr (refl::is_reflectable<TBlock>()) {
-            std::lock_guard lg(_lock);
+            std::lock_guard lg(_mutex);
 
             // prepare old settings if required
             property_map oldSettings;
             if constexpr (HasSettingsChangedCallback<TBlock>) {
                 storeDefaultSettings(oldSettings);
             }
+            if (currentTime == 0ULL) {
+                currentTime = settings::convertTimePointToUint64Ns(std::chrono::system_clock::now());
+            }
+
+            const SettingsCtx  activeSettingsCtx{currentTime, _activeCtx};
+            const property_map bestMatchParameters = getBestMatchStoredParameters(activeSettingsCtx);
 
             // check if reset of settings should be performed
-            if (_staged.contains(gr::tag::RESET_DEFAULTS)) {
-                _staged.clear();
+            if (bestMatchParameters.contains(gr::tag::RESET_DEFAULTS)) {
                 resetDefaults();
             }
 
             // update staged and forward parameters based on member properties
             property_map staged;
-            for (const auto& [key, stagedValue] : _staged) {
-                auto applyOneMemberChanges = [&key, &staged, &result, &stagedValue, this](auto member) {
-                    using RawType = std::remove_cvref_t<decltype(member(*_block))>;
+            for (const auto& [key, stagedValue] : bestMatchParameters) {
+                auto applyOneMemberChanges = [&key, &staged, &result, &stagedValue, this]<typename TFieldMeta>(TFieldMeta member) {
+                    using RawType = std::remove_cvref_t<typename TFieldMeta::value_type>;
                     using Type    = unwrap_if_wrapped_t<RawType>;
                     if constexpr (settings::isWritableMember<Type>(member)) {
                         if (std::string(get_display_name(member)) == key && std::holds_alternative<Type>(stagedValue)) {
                             if constexpr (is_annotated<RawType>()) {
                                 if (member(*_block).validate_and_set(std::get<Type>(stagedValue))) {
-                                    result.appliedParameters.insert_or_assign(key, stagedValue);
                                     if constexpr (HasSettingsChangedCallback<TBlock>) {
                                         staged.insert_or_assign(key, stagedValue);
                                     } else {
@@ -589,7 +594,7 @@ public:
                                         RawType::LimitType::ValidatorFunc == nullptr ? "not" : "");
 #else
                                     fmt::print(stderr, " cannot set field {}({})::{} = {} to {} due to limit constraints [{}, {}] validate func is {} defined\n", //
-                                        "_block->unique_name", "_block->name", "member(*_block)", std::get<Type>(stagedValue),                                    //
+                                        "_block->uniqueName", "_block->name", member(*_block), std::get<Type>(stagedValue),                                       //
                                         std::string(get_display_name(member)), RawType::LimitType::MinRange,
                                         RawType::LimitType::MaxRange, //
                                         RawType::LimitType::ValidatorFunc == nullptr ? "not" : "");
@@ -605,7 +610,7 @@ public:
                                 }
                             }
                         }
-                        if (_auto_forward.contains(key)) {
+                        if (_autoForwardParameters.contains(key)) {
                             result.forwardParameters.insert_or_assign(key, stagedValue);
                         }
                     }
@@ -614,47 +619,46 @@ public:
             }
 
             // update active parameters
-            auto update_active = [this](auto member) {
-                using Type = unwrap_if_wrapped_t<std::remove_cvref_t<decltype(member(*_block))>>;
+            auto updateActive = [this]<typename TFieldMeta>(TFieldMeta member) {
+                using Type = unwrap_if_wrapped_t<std::remove_cvref_t<typename TFieldMeta::value_type>>;
                 if constexpr (settings::isReadableMember<Type>(member)) {
-                    _active.insert_or_assign(get_display_name(member), static_cast<Type>(member(*_block)));
+                    _activeParameters.insert_or_assign(get_display_name(member), static_cast<Type>(member(*_block)));
                 }
             };
-            processMembers<TBlock>(update_active);
+            processMembers<TBlock>(updateActive);
 
             // invoke user-callback function if staged is not empty
             if (!staged.empty()) {
-                if constexpr (requires { _block->settingsChanged(/* old settings */ _active, /* new settings */ staged); }) {
+                if constexpr (requires { _block->settingsChanged(/* old settings */ _activeParameters, /* new settings */ staged); }) {
                     _block->settingsChanged(/* old settings */ oldSettings, /* new settings */ staged);
-                } else if constexpr (requires { _block->settingsChanged(/* old settings */ _active, /* new settings */ staged, /* new forward settings */ result.forwardParameters); }) {
+                } else if constexpr (requires { _block->settingsChanged(/* old settings */ _activeParameters, /* new settings */ staged, /* new forward settings */ result.forwardParameters); }) {
                     _block->settingsChanged(/* old settings */ oldSettings, /* new settings */ staged, /* new forward settings */ result.forwardParameters);
                 }
             }
 
-            if (_staged.contains(gr::tag::STORE_DEFAULTS)) {
+            if (bestMatchParameters.contains(gr::tag::STORE_DEFAULTS)) {
                 storeDefaults();
             }
 
             if constexpr (HasSettingsResetCallback<TBlock>) {
-                if (_staged.contains(gr::tag::RESET_DEFAULTS)) {
+                if (bestMatchParameters.contains(gr::tag::RESET_DEFAULTS)) {
                     _block->reset();
                 }
             }
-
-            _staged.clear();
+            removeExpiredStoredParameters(activeSettingsCtx);
         }
 
-        SettingsBase::_changed.store(false);
+        _changed.store(false);
         return result;
     }
 
     void updateActiveParameters() noexcept override {
         if constexpr (refl::is_reflectable<TBlock>()) {
-            std::lock_guard lg(_lock);
-            auto            processOneMember = [&, this](auto member) {
-                using Type = unwrap_if_wrapped_t<std::remove_cvref_t<decltype(member(*_block))>>;
+            std::lock_guard lg(_mutex);
+            auto            processOneMember = [&, this]<typename TFieldMeta>(TFieldMeta member) {
+                using Type = unwrap_if_wrapped_t<std::remove_cvref_t<typename TFieldMeta::value_type>>;
                 if constexpr (settings::isReadableMember<Type>(member)) {
-                    _active.insert_or_assign(get_display_name(member), static_cast<Type>(member(*_block)));
+                    _activeParameters.insert_or_assign(get_display_name_const(member).str(), member(*_block));
                 }
             };
             processMembers<TBlock>(processOneMember);
@@ -662,11 +666,153 @@ public:
     }
 
 private:
+    [[nodiscard]] std::optional<pmtv::pmt> findBestMatchCtx(const pmtv::pmt& contextToSearch) const {
+        if (_storedParameters.empty()) {
+            return std::nullopt;
+        }
+
+        // exact match
+        if (_storedParameters.find(contextToSearch) != _storedParameters.end()) {
+            return contextToSearch;
+        }
+
+        // retry until we either get a match or std::nullopt
+        for (std::size_t attempt = 0;; ++attempt) {
+            for (const auto& i : _storedParameters) {
+                const auto matches = _matchPred(i.first, contextToSearch, attempt);
+                if (!matches) {
+                    return std::nullopt;
+                } else if (*matches) {
+                    return i.first; // return the best matched SettingsCtx.context
+                }
+            }
+        }
+        return std::nullopt;
+    }
+
+    [[nodiscard]] bool hasStoredParameters(const SettingsCtx& ctx) const {
+        const auto bestMatchCtx = findBestMatchCtx(ctx.context);
+        if (bestMatchCtx == std::nullopt) {
+            return false;
+        }
+        return !_storedParameters[bestMatchCtx.value()].empty();
+    }
+
+    [[nodiscard]] property_map getBestMatchStoredParameters(const SettingsCtx& ctx) const {
+        const auto bestMatchCtx = findBestMatchCtx(ctx.context);
+        if (bestMatchCtx == std::nullopt) {
+            return property_map();
+        }
+        const auto& vec = _storedParameters[bestMatchCtx.value()];
+        if (vec.empty()) {
+            return property_map();
+        }
+        if (ctx.time == 0ULL || vec.back().first.time <= ctx.time) {
+            return vec.back().second;
+        } else {
+            auto lower = std::ranges::lower_bound(vec, ctx.time, {}, [](const auto& a) { return a.first.time; });
+            if (lower == vec.end()) {
+                return vec.back().second;
+            } else {
+                if (lower->first.time == ctx.time) {
+                    return lower->second;
+                } else if (lower != vec.begin()) {
+                    --lower;
+                    return lower->second;
+                }
+            }
+        }
+        return property_map();
+    }
+
+    void addStoredParameters(const property_map& newParameters, const SettingsCtx& ctx) {
+        _storedParameters[ctx.context].push_back({ctx, newParameters});
+        auto& vec = _storedParameters[ctx.context];
+        std::sort(vec.begin(), vec.end(), [](const auto& a, const auto& b) { return a.first.time < b.first.time; });
+
+        if (_autoUpdateParameters.find(ctx.context) == _autoUpdateParameters.end()) {
+            _autoUpdateParameters[ctx.context] = _allWritableMembers;
+        }
+    }
+
+    void removeExpiredStoredParameters(const SettingsCtx& ctx) {
+        const auto bestMatchCtx = findBestMatchCtx(ctx.context);
+        if (bestMatchCtx == std::nullopt) {
+            return;
+        }
+        auto& vec = _storedParameters[bestMatchCtx.value()];
+        if (vec.empty()) {
+            return;
+        }
+        if (ctx.time == 0ULL || vec.back().first.time <= ctx.time) {
+            vec.clear();
+        } else {
+            auto lower = std::ranges::lower_bound(vec, ctx.time, {}, [](const auto& a) { return a.first.time; });
+            if (lower == vec.end()) {
+                vec.clear();
+            } else {
+                if (lower->first.time == ctx.time) {
+                    vec.erase(vec.begin(), lower + 1);
+                } else if (lower != vec.begin()) {
+                    vec.erase(vec.begin(), lower);
+                }
+            }
+        }
+    }
+
+    [[nodiscard]] bool isContextPresentInTag(const Tag& tag) const {
+        if (tag.map.contains(gr::tag::TRIGGER_META_INFO.shortKey())) {
+            const pmtv::pmt& pmtMetaInfo = tag.map.at(std::string(gr::tag::TRIGGER_META_INFO.shortKey()));
+            if (std::holds_alternative<property_map>(pmtMetaInfo)) {
+                const property_map& metaInfo = std::get<property_map>(pmtMetaInfo);
+                if (metaInfo.contains(std::string(gr::tag::CONTEXT.shortKey()))) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    [[nodiscard]] bool isTriggeredTimePresentInTag(const Tag& tag) const {
+        if (tag.map.contains(gr::tag::TRIGGER_TIME.shortKey())) {
+            const pmtv::pmt& pmtTimeUtcNs = tag.map.at(std::string(gr::tag::TRIGGER_TIME.shortKey()));
+            if (std::holds_alternative<uint64_t>(pmtTimeUtcNs)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    [[nodiscard]] SettingsCtx createSettingsCtxFromTag(const Tag& tag) const {
+        // If TRIGGER_META_INFO is not present then context =_activeCtx, time = now()
+        // If CONTEXT is not present then context =_activeCtx
+        // IF TRIGGER_TIME is not present then time = now()
+
+        SettingsCtx ctx(0ULL, _activeCtx);
+
+        // update if context is present
+        if (isContextPresentInTag(tag)) {
+            const pmtv::pmt& pmtMetaInfo = tag.map.at(std::string(gr::tag::TRIGGER_META_INFO.shortKey()));
+            ctx.context                  = std::get<property_map>(pmtMetaInfo).at(std::string(gr::tag::CONTEXT.shortKey()));
+        }
+
+        // update trigger time if present
+        if (isTriggeredTimePresentInTag(tag)) {
+            ctx.time = std::get<uint64_t>(tag.map.at(std::string(gr::tag::TRIGGER_TIME.shortKey())));
+        }
+
+        if (ctx.time == 0ULL) {
+            ctx.time = settings::convertTimePointToUint64Ns(std::chrono::system_clock::now());
+        }
+        return ctx;
+    }
+
     void storeDefaultSettings(property_map& oldSettings) {
         // take a copy of the field -> map value of the old settings
         if constexpr (refl::is_reflectable<TBlock>()) {
-            auto processOneMember = [&, this](auto member) {
-                using Type = unwrap_if_wrapped_t<std::remove_cvref_t<decltype(member(*_block))>>;
+            auto processOneMember = [&, this]<typename TFieldMeta>(TFieldMeta member) {
+                using Type = unwrap_if_wrapped_t<std::remove_cvref_t<typename TFieldMeta::value_type>>;
+
                 if constexpr (settings::isReadableMember<Type>(member)) {
                     oldSettings.insert_or_assign(get_display_name(member), pmtv::pmt(member(*_block)));
                 }
@@ -682,9 +828,7 @@ private:
         }
         refl::util::for_each(refl::reflect<T>().members, func);
     }
-};
-
-static_assert(SettingsLike<BasicSettings<int>>);
+}; // class CtxSettings
 
 } // namespace gr
